+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,8 +1,5 @@
 package roomescape.controller;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.List;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,50 +10,29 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.controller.dto.ReservationRequest;
 import roomescape.controller.dto.ReservationResponse;
-import roomescape.domain.Reservation;
-import roomescape.repository.ReservationRepository;
+import roomescape.service.ReservationService;
 
 @RestController
 @RequestMapping("/reservations")
 public class ReservationController {
-    private final ReservationRepository reservationRepository;
+    private final ReservationService reservationService;
 
-    public ReservationController(ReservationRepository reservationRepository) {
-        this.reservationRepository = reservationRepository;
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @PostMapping
-    public ReservationResponse saveReservation(@RequestBody ReservationRequest reservationRequest) {
-        Reservation reservation = fromRequest(reservationRequest);
-        Reservation savedReservation = reservationRepository.saveReservation(reservation);
-        return toResponse(savedReservation);
-    }
-
-    private Reservation fromRequest(ReservationRequest reservationRequest) {
-        String name = reservationRequest.name();
-        LocalDate date = reservationRequest.date();
-        LocalTime time = reservationRequest.time();
-        LocalDateTime dateTime = LocalDateTime.of(date, time);
-        return new Reservation(name, dateTime);
-    }
-
-    private ReservationResponse toResponse(Reservation reservation) {
-        return new ReservationResponse(reservation.getId(),
-                reservation.getName(), reservation.getDate(), reservation.getTime());
+    public ReservationResponse reservationSave(@RequestBody ReservationRequest reservationRequest) {
+        return reservationService.addReservation(reservationRequest);
     }
 
     @GetMapping
-    public List<ReservationResponse> findAllReservations() {
-        List<Reservation> reservations = reservationRepository.findAllReservation();
-
-        return reservations.stream()
-                .map(reservation -> new ReservationResponse(reservation.getId(), reservation.getName(),
-                        reservation.getDate(), reservation.getTime()))
-                .toList();
+    public List<ReservationResponse> reservaionList() {
+        return reservationService.findReservations();
     }
 
     @DeleteMapping("/{reservationId}")
-    public void delete(@PathVariable long reservationId) {
-        reservationRepository.deleteReservationById(reservationId);
+    public void reservationRemove(@PathVariable long reservationId) {
+        reservationService.removeReservations(reservationId);
     }
 }
